dumped a replica taxon tree into sqllite database, completed test_sql_tools, started work on tests for taxon tree importer methods
