+++ tests/test_sql_tools.py
@@ -1,7 +1,6 @@
 """This file contains unit tests for picturae_import.py"""
 import unittest
 import logging
-import sqlite3
 import numpy as np
 import pandas as pd
 
@@ -9,22 +8,9 @@ from image_client.sql_csv_utils import insert_table_record
 from tests.pic_importer_test_class import TestPicturaeImporter
 import image_client.sql_csv_utils as scu
 from tests.testing_tools import TestingTools
-from image_client.casbotany_sql_lite import *
+from casbotany_sql_lite import *
 from image_client import time_utils
 from uuid import uuid4
-#
-# class TestSQLlite(unittest.TestCase):
-#     def test_casbotanylite(self):
-#         connection = sqlite3.connect('cas_botanylite.db')
-#         curs = connection.cursor()
-#         curs.execute('''SELECT * FROM agent''')
-#         num_columns = len(curs.description)
-#         self.assertEqual(num_columns, 45)
-#         curs.close()
-#         connection.close()
-#
-# # class for testing csv_import function
-# # too many redundant t4e
 
 
 ## this one should go to a test file for sql_db_utils
@@ -41,31 +27,14 @@ class TestSqlInsert(unittest.TestCase, TestingTools):
                                                            paths=self.md5_hash)
 
     def test_casbotanylite(self):
-        "testing the whether the sqllite datbase can connect"
-        connection = sqlite3.connect('cas_botanylite.db')
+        "testing the wether the sqllite datbase can connect"
+        connection = sqlite3.connect('casbotany_lite.db')
         curs = connection.cursor()
         curs.execute('''SELECT * FROM agent''')
         num_columns = len(curs.description)
         self.assertEqual(num_columns, 45)
         curs.close()
         connection.close()
-
-
-    def test_sql_populate(self):
-        """testing if the joining sql_operations in the
-           function populate are creating the right sql string"""
-        table_select = 'agent'
-        column_list = ['AgentType', 'FirstName', 'MiddleInitial']
-        value_list = [1, 'Fake', 'Name']
-
-        column_list = ', '.join(column_list)
-        value_list = ', '.join(f"'{value}'" if isinstance(value, str) else repr(value) for value in value_list)
-
-        sql = f'''INSERT INTO {table_select} ({column_list}) VALUES({value_list})'''
-        # assert statement
-        expected_output = f'''INSERT INTO agent (AgentType, FirstName, MiddleInitial) VALUES(1, 'Fake', 'Name')'''
-        self.assertEqual(sql, expected_output)
-
 
     def test_sql_string(self):
         """testing if create_sql_string
@@ -81,6 +50,19 @@ class TestSqlInsert(unittest.TestCase, TestingTools):
 
         self.assertEqual(sql, f'''INSERT INTO cattab (tax1, tax2, tax3, feline1) VALUES(1, 2, 3, 'cat');''')
 
+    def test_remove_two_index(self):
+        """tests whether remove two index will drop
+            correct amount of terms from value/column lists"""
+        test_values = [1, 2, pd.NA, "cat", '', np.nan, None]
+        test_col = ["col1", "col2", "col3", "col4", "col5", "col6", "col7"]
+
+        test_values, test_col = scu.remove_two_index(value_list=test_values, column_list=test_col)
+
+        self.assertEqual(len(test_col), len(test_values))
+        self.assertEqual(len(test_values), 3)
+        # using pd.isna() to avoid "boolean value of NA is ambiguous error
+        for term in test_values:
+            self.assertTrue(not pd.isna(term) and term not in [None, np.nan, ''])
 
     def test_create_locality(self):
         """testing create_locality function by
@@ -158,7 +140,7 @@ class TestSqlInsert(unittest.TestCase, TestingTools):
         # removing na values from both lists
         value_list, column_list = scu.remove_two_index(value_list, column_list)
 
-        # assert that len val list and column list are equevalent.
+        # assert that len val list and column list are equivalent.
 
         self.assertEqual(len(value_list), len(column_list))
 
@@ -167,28 +149,12 @@ class TestSqlInsert(unittest.TestCase, TestingTools):
 
         insert_table_record(connection=self.connection, logger_int=self.logger, sql=sql, sqllite=True)
 
-
         station_field = casbotany_lite_getrecord(f'''SELECT `StationFieldNumber` FROM collectingevent WHERE 
-                                                            `StationFieldNumber` = {123456}''')
+                                                     `StationFieldNumber` = {123456}''')
 
         # asserting that station field number is in right column
 
         self.assertEqual('123456', station_field)
-
-
-    def test_remove_two_index(self):
-        """tests whether remove two index will drop
-            correct amount of terms from value/column lists"""
-        test_values = [1, 2, pd.NA, "cat", '', np.nan, None]
-        test_col = ["col1", "col2", "col3", "col4", "col5", "col6", "col7"]
-
-        test_values, test_col = scu.remove_two_index(value_list=test_values, column_list=test_col)
-
-        self.assertEqual(len(test_col), len(test_values))
-        self.assertEqual(len(test_values), 3)
-        # using pd.isna() to avoid "boolean value of NA is ambiguous error
-        for term in test_values:
-            self.assertTrue(not pd.isna(term) and term not in [None, np.nan, ''])
 
     def tearDown(self):
         """deleting instance of PicturaeImporter"""
